Twee CRUD interfaces
twee CRUD interfaces geimplementeerd

+++ controllers/categoryController.js
@@ -2,13 +2,36 @@ const db = require('../config/db');
 
 
 const getCategories = (req, res) => {
-  db.query('SELECT * FROM category', (err, results) => {
+  const limit = parseInt(req.query.limit) || 1000;
+  const offset = parseInt(req.query.offset) || 0;
+  const search = req.query.search || '';  
+
+  
+  const query = 'SELECT * FROM category WHERE name LIKE ? LIMIT ? OFFSET ?';
+  db.query(query, [`%${search}%`, limit, offset], (err, results) => {
     if (err) {
       return res.status(500).send('Error retrieving categories');
     }
     res.json(results);
   });
 };
+
+
+const getCategoryById = (req, res) => {
+  const { id } = req.params;
+
+  const query = 'SELECT * FROM category WHERE id = ?';
+  db.query(query, [id], (err, result) => {
+    if (err) {
+      return res.status(500).send('Error retrieving category');
+    }
+    if (result.length === 0) {
+      return res.status(404).send('Category not found');
+    }
+    res.json(result[0]);
+  });
+};
+
 
 
 const createCategory = (req, res) => {
@@ -27,8 +50,15 @@ const createCategory = (req, res) => {
   });
 };
 
+
 const updateCategory = (req, res) => {
   const { id } = req.params;
+
+ 
+  if (isNaN(id)) {
+    return res.status(400).send('ID moet een nummer zijn.');
+  }
+
   const { name, description } = req.body;
 
   const query = 'UPDATE category SET name = ?, description = ? WHERE id = ?';
@@ -44,13 +74,51 @@ const updateCategory = (req, res) => {
 const deleteCategory = (req, res) => {
   const { id } = req.params;
 
-  const query = 'DELETE FROM category WHERE id = ?';
-  db.query(query, [id], (err, result) => {
+  
+  if (isNaN(id)) {
+    return res.status(400).send('ID moet een nummer zijn.');
+  }
+
+  
+  const checkQuery = 'SELECT * FROM category WHERE id = ?';
+  db.query(checkQuery, [id], (err, result) => {
     if (err) {
-      return res.status(500).send('Error deleting category');
+      return res.status(500).send('Error checking category');
     }
-    res.send('Category deleted');
+
+    
+    if (result.length === 0) {
+      return res.status(404).send('Category not found');
+    }
+
+    
+    const deleteQuery = 'DELETE FROM category WHERE id = ?';
+    db.query(deleteQuery, [id], (err, result) => {
+      if (err) {
+        return res.status(500).send('Error deleting category');
+      }
+      res.send('Category deleted');
+    });
   });
 };
 
-module.exports = { getCategories, createCategory, updateCategory, deleteCategory };
+
+
+const searchCategories = (req, res) => {
+  const search = req.query.search || '';  
+
+  const query = 'SELECT * FROM category WHERE name LIKE ?';
+  
+  db.query(query, [`%${search}%`], (err, results) => {
+    if (err) {
+      return res.status(500).send('Error retrieving categories');
+    }
+    if (results.length === 0) {
+      return res.status(404).send('Category not found');
+    }
+    res.json(results);  
+  });
+};
+
+
+module.exports = { getCategories, createCategory, updateCategory, deleteCategory, getCategoryById, searchCategories };
